% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_for_esa.R
\name{clean_for_esa}
\alias{clean_for_esa}
\title{Prepare an ESA Online Logbook Entry}
\usage{
clean_for_esa(.df)
}
\description{
`clean_for_esa()` attaches building site numbers to the Megamation
export "CP_ESA_WO_REPORT" of the WORK_ORDER_MASTER.
Data is then cleaned for upload to the Electrical Safety Authority
(ESA) Online website.
}
