% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaApp-package.R
\docType{package}
\name{esaApp-package}
\alias{esaApp}
\alias{esaApp-package}
\title{esaApp: Data Converter: Megamation to ESA}
\description{
Attaches building site numbers to the Megamation export 'CP_ESA_WO_REPORT' of the WORK_ORDER_MASTER which is then cleaned for upload to the Electrical Safety Authority (ESA) Online website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asadow/esaApp}
  \item Report bugs at \url{https://github.com/asadow/esaApp/issues}
}

}
\author{
\strong{Maintainer}: Adam Sadowski \email{asadowsk@uoguelph.ca}

}
\keyword{internal}
