% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clean_for_esa.R
\name{fct_clean_for_esa}
\alias{fct_clean_for_esa}
\title{Prepare an ESA Online Logbook Entry}
\usage{
fct_clean_for_esa(.data)
}
\arguments{
\item{.data}{A data frame.}
}
\description{
`clean_for_esa()` cleans Megamation exports for upload to the
Electrical Safety Authority (ESA) Online website.
Requirements are stated in the LogbookEntryUploadForm.xls.
}
