% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitenumber}
\alias{sitenumber}
\title{Dictionary: Building Identifiers}
\format{
`sitenumber`: A data frame
 with `r nrow(sitenumber)` rows and `r ncol(sitenumber)` columns:
\describe{
  \item{site_number}{Site number.}
  \item{bldg_no}{Building number.}
  }
}
\source{
Raw data is the file named
"CSSL Site Listing with Site Numbers - as of October 5 2022.xls",
sent directly to Chris Payne from Sherry at ESA.
Chris did not have credentials to log into the ESA reporting site,
so he simply asked for the info.
}
\usage{
sitenumber
}
\description{
Names, numbers, and addresses for buildings and sites.
}
\examples{
sitenumber
}
\keyword{datasets}
