% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{utils_date_range}
\alias{utils_date_range}
\title{Get date range from a Megamation export}
\usage{
utils_date_range(path)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}
}
\description{
The date range in a Megamation export is contained in the first three lines.
}
