% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{utils_DT}
\alias{utils_DT}
\title{Table-only `datatable()` that fills container and has variable rows.}
\usage{
utils_DT(.data, rows)
}
\arguments{
\item{.data}{A data frame.}

\item{rows}{Number of rows.}
}
\description{
`fillContainer = TRUE` is needed for a scrollable container
in `card_body_fill()`
}
