% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{utils_load_xl}
\alias{utils_load_xl}
\title{Load an Excel file}
\usage{
utils_load_xl(path, name, ...)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{name}{Name of the file.}

\item{...}{Arguments to pass on to `readxl::read_excel()`}
}
\description{
Load an Excel file
}
